Started moving vector parts into separate files
